Rename script, use readlink, add main logic (not tested)
Since creation date is not supported by all filesystems in linux,
script will look at and delete files according to modification
date instead. Hence script will be renamed 'cdate...'->'mdate..."

Assigning variables relating to file names and path with readlink
instead of dirname (not always accurate) or realpath (deprecated).

Main if conditionals for checking if name is file (not dir), if
file was last modified earlier than a certain time, and removing
that file was added.

This commit has not yet been tested.